Make search term coloring more deterministic
If a matched text has, for instance a red color, setting search
highlight to reverse makes the text exhibiting a red color background,
which can be confusing especially when search matches multiple texts,
each with different color.

Setting the foreground and background explicitly so as search term has a
more uniform and deterministic highlighting.